+++ STC32F_Code/Project/USER/inc/magnet.h
@@ -14,7 +14,7 @@ void Magnet_ADC_Read(void);
 void Inductor_Process(void);
 void Magnet_ADC_Filter(void);
 void Bubble_Sort_ADC(void);
-void Bubble_Sort_Int(int data_array[], int length);
+void Bubble_Sort_Int(int16 data_array[], int16 length);
 void Inductor_Normal(void);
 void Inductor_Analyse(void);
