+++ STC32F_Code/Project/USER/inc/motor.h
@@ -7,7 +7,7 @@ extern int16 motor_left_pwm, motor_right_pwm;
 
 void Motor_PWM_Init(void);
 void Motor_PWM_Write(void);
-int Limit(int x, int min, int max);
+int16 Limit(int16 x, int16 min, int16 max);
 
 
 #endif
