+++ STC32F_Code/Project/USER/src/beep.c
@@ -7,7 +7,7 @@ void BEEP_Init(void)
     gpio_mode(P3_2, GPO_PP); // ·äÃùÆ÷
 }
 
-void BEEP_ON_ms(int ms)
+void BEEP_ON_ms(int16 ms)
 {
     BEEP = 1;
     delay_ms(ms);
