+++ STC32F_Code/Project/USER/src/pid.c
@@ -5,9 +5,9 @@ PIDTypeDef direction, motor_left, motor_right;
 float direction_output;
 int16 position_last;
 
-int motor_left_error, motor_right_error;
-int motor_left_last_error, motor_right_last_error;
-int target_speed_left, target_speed_right;
+int16 motor_left_error, motor_right_error;
+int16 motor_left_last_error, motor_right_last_error;
+int16 target_speed_left, target_speed_right;
 
 void PID_Parameter_Init(PIDTypeDef *sptr, float KP, float KI, float KD, float KP_2, float KD_2)
 {
@@ -49,7 +49,7 @@ void Direction_PID(void)
 // 左轮内环
 void Left_Speed_PID(void)
 {
-    motor_left_error = (int)(target_speed_left - encoder_left.encoder_now);
+    motor_left_error = (int16)(target_speed_left - encoder_left.encoder_now);
     motor_left_pwm += (motor_left_error - motor_left_last_error) * motor_left.KP + motor_left_error * motor_left.KI;
     motor_left_last_error = motor_left_error;
 
@@ -60,7 +60,7 @@ void Left_Speed_PID(void)
 // 右轮内环
 void Right_Speed_PID(void)
 {
-    motor_right_error = (int)(target_speed_right - encoder_right.encoder_now);
+    motor_right_error = (int16)(target_speed_right - encoder_right.encoder_now);
     motor_right_pwm += (motor_right_error - motor_right_last_error) * motor_right.KP + motor_right_error * motor_left.KI;
     motor_right_last_error = motor_right_error;
 
@@ -68,7 +68,7 @@ void Right_Speed_PID(void)
     motor_right_pwm = clamp(motor_right_pwm, -3500, 8000); //PWM限幅
 }
 
-int clamp(int value, int min_value, int max_value)
+int16 clamp(int16 value, int16 min_value, int16 max_value)
 {
     if (value < min_value)
     {
